Made just a compute class for inheriting
